// src/app/Core/Interfaces/ILegalRepresentative.ts
export interface ILegalRepresentative {
    Name: string;
    Nationality?: string;
    Ocupation?: string;
    MaritalStatus?: string;
    GeneralId?: string;
    NationalId?: string;
    ZipCode?: string;
    Address?: string;
    Complement?: string;
    Number?: number;
    Suburb?: string;
    City?: string;
    State?: string;
    AreaCode?: string;
    PhoneNumber?: string;
    CellPhone?: string;  
  }